# docs/swagger.json
{ "openapi": "3.0.0",
    "info":{
        "title":"Speckio Web App",
        "description":"This is an express API for the Speckio web app. Made by NHMG Development.",
        "version":"1.0.0"
    },
    "servers":[
        {
            "url": "https://localhost"
        }
    ],
    "components": {
      "securitySchemes": {
        "ApiKeyAuth": {
          "type": "apiKey",
          "in": "header",
          "name": "authorization"
        }
      }
    },
    "security": [
      { "ApiKeyAuth": [] }
    ],
    "paths":{
...
      "/user": {
        "get": {
          "tags": [
            "User"
          ],
          "description": "Return details of current user.",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "user": {
                        "type": "object",
                        "example": {
                          "firstName": "John",
                          "lastName": "Smith",
                          "email": "johns@speckio.com"
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "User not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "patch": {
          "tags": [
            "User"
          ],
          "description": "Update details of current user.",
          "parameters": [
            {
              "name": "firstName",
              "in": "query",
              "description": "First name of user",
              "required": false,
              "schema": {
                "type": "string",
                "example": "John"
              }
            },
            {
              "name": "lastName",
              "in": "query",
              "description": "Last name of user",
              "required": false,
              "schema": {
                "type": "string",
                "example": "Smith"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "User info updated"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "User not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/user/team": {
        "get": {
          "tags": [
            "User"
          ],
          "description": "Return an array of teams for a given user.",
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "teams": {
                        "type": "array",
                        "example": [
                          {
                            "teamID": 1,
                            "teamName": "Speckio Development Team",
                            "teamDescription": "Web development team at Speckio.",
                            "teamLeaderID": 2,
                            "firstName": "Jane",
                            "lastName": "Doe",
                            "hasAccepted": 1
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "No teams found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "User"
          ],
          "description": "Delete current user from a given team",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team",
              "required": true,
              "schema": {
                "type": "integer",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "User removed from team"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "No teams found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "/team": {
        "get": {
          "tags": [
            "Team"
          ],
          "description": "Return details of given team",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team to search for",
              "required": true,
              "schema": {
                "type": "number",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "team": {
                        "type": "object",
                        "example": {
                          "teamID": 1,
                          "teamName": "Speckio Development Team",
                          "teamDescription": "Web development team at Speckio.",
                          "teamLeaderID": 2,
                          "firstName" : "Jane",
                          "lastName" : "Doe"                          
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Team not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Team"
          ],
          "description": "Create a new team",
          "parameters": [
            {
              "name": "name",
              "in": "query",
              "description": "Name of team",
              "required": true,
              "schema": {
                "type": "string",
                "example": "NHMG Development"
              }
            },
            {
              "name": "desc",
              "in": "query",
              "description": "Description of team",
              "required": true,
              "schema": {
                "type": "string",
                "example": "QUT Capstone Team"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "team created"
                      },
                      "teamID": {
                        "type": "integer",
                        "example": "2"
                      }    
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Team not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "patch": {
          "tags": [
            "Team"
          ],
          "description": "Edit a team",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team",
              "required": true,
              "schema": {
                "type": "integer",
                "example": 2
              }
            },
            {
              "name": "name",
              "in": "query",
              "description": "Name of team",
              "required": true,
              "schema": {
                "type": "string",
                "example": "NHMG Development 2"
              }
            },
            {
              "name": "desc",
              "in": "query",
              "description": "Description of team",
              "required": true,
              "schema": {
                "type": "string",
                "example": "QUT Capstone Team 3"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "team created"
                      },
                      "teamID": {
                        "type": "integer",
                        "example": "2"
                      }    
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Team not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Team"
          ],
          "description": "Delete a given team.",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team to delete",
              "required": true,
              "schema": {
                "type": "number",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "team": {
                        "type": "object",
                        "example": {
                          "message": "Team removed"                        
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Team not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/team/overview": {
        "get": {
          "tags": [
            "Team"
          ],
          "description": "Return an array of users strengths and weaknesses in a given team.",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team to search for",
              "required": true,
              "schema": {
                "type": "number",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "overview": {
                        "type": "array",
                        "example": [
                          {
                            "userID": 1,
                            "firstName": "John",
                            "lastName": "Smith",
                            "descName": "High Interpersonal Relationships",
                            "strengths": "High Interpersonal Relationships Strengths",
                            "weaknesses": "High Interpersonal Relationships Weaknesses"
                          },
                          {
                            "userID": 2,
                            "firstName": "Jane",
                            "lastName": "Doe",
                            "descName": "Mid Interpersonal Relationships",
                            "strengths": "Mid Interpersonal Relationships Strengths",
                            "weaknesses": "Mid Interpersonal Relationships Weaknesses"
                          },
                          {
                            "userID": 3,
                            "firstName": "George",
                            "lastName": "Adams",
                            "descName": "Mid Interpersonal Relationships",
                            "strengths": "Mid Interpersonal Relationships Strengths",
                            "weaknesses": "Mid Interpersonal Relationships Weaknesses"
                          },
                          {
                            "userID": 4,
                            "firstName": "Lucy",
                            "lastName": "Barnes",
                            "descName": "Low Interpersonal Relationships",
                            "strengths": "Low Interpersonal Relationships Strengths",
                            "weaknesses": "Low Interpersonal Relationships Weaknesses"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Team not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/team/user": {
        "get": {
          "tags": [
            "Team"
          ],
          "description": "Return an array of users in a given team.",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team to search for",
              "required": true,
              "schema": {
                "type": "number",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "team": {
                        "type": "array",
                        "example": [
                          {
                            "userID": 1,
                            "firstname": "John",
                            "lastname": "Smith",
                            "email": "johns@speckio.com"
                          },
                          {
                            "userID": 2,
                            "firstname": "Jane",
                            "lastname": "Doe",
                            "email": "janed@speckio.com"
                          },
                          {
                            "userID": 3,
                            "firstname": "George",
                            "lastname": "Adams",
                            "email": "georgea@speckio.com"
                          },
                          {
                            "userID": 4,
                            "firstname": "Lucy",
                            "lastname": "Barnes",
                            "email": "lucyb@speckio.com"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Team not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Team"
          ],
          "description": "Add user to team",
          "parameters": [
            {
              "name": "email",
              "in": "query",
              "description": "Email of user",
              "required": true,
              "schema": {
                "type": "string",
                "example": "walterf@speckio.com"
              }
            },
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team",
              "required": true,
              "schema": {
                "type": "number",
                "example": "1"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "team": {
                        "type": "array",
                        "example": [
                          {"firstname":"John","lastname":"Smith","email":"johns@speckio.com"},
                          {"firstname":"Jane","lastname":"Doe","email":"janed@speckio.com"},
                          {"firstname":"George","lastname":"Adams","email":"georgea@speckio.com"},
                          {"firstname":"Lucy","lastname":"Barnes","email":"lucyb@speckio.com"}
                        ]
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Team not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "Team"
          ],
          "description": "Delete a user from a team",
          "parameters": [
            {
              "name": "teamID",
              "in": "query",
              "description": "ID of team",
              "required": true,
              "schema": {
                "type": "number",
                "example": 1
              }
            },
            {
              "name": "userID",
              "in": "query",
              "description": "ID of user",
              "required": true,
              "schema": {
                "type": "number",
                "example": 1
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "team": {
                        "type": "object",
                        "example": {
                          "message": "User removed from team"                        
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Bad Request",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "messege": {
                        "type": "string",
                        "example": "Teammember not found"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
}